requirements.txt
Django==4.2.1
djangorestframework
djangorestframework-simplejwt
psycopg2-binary